all: migrate back to Arduino IDE
For some reason the same code while being compiled from PlatformIO
seems to break the USB stack in a way that the Pico is not being
detected in Windows neither as HID nor as MIDI device. It still
works in Linux, but breaks on Windows, and the same code compiled
from within Arduino IDE works in both. I am not sure what the problem
is, and I dont have time right now to file an issue, and I need this
to work in Windows for Lightroom automation, so I had to migrate it
back.